Move C data flow tests into fuzzyc tests
