+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/dotgenerator/DotDdgGeneratorTests.scala
@@ -1,8 +1,8 @@
-package io.shiftleft.dataflowengineoss.dotgenerator
+package io.shiftleft.fuzzyc2cpg.dotgenerator
 
-import io.shiftleft.dataflowengineoss.language.DataFlowCodeToCpgSuite
-import io.shiftleft.semanticcpg.language._
 import io.shiftleft.dataflowengineoss.language._
+import io.shiftleft.fuzzyc2cpg.testfixtures.DataFlowCodeToCpgSuite
+import io.shiftleft.semanticcpg.language._
 
 class DotDdgGeneratorTests extends DataFlowCodeToCpgSuite {
   override val code =
