+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/querying/CDataFlowTests.scala
@@ -1,8 +1,10 @@
-package io.shiftleft.dataflowengineoss.language
+package io.shiftleft.fuzzyc2cpg.querying
 
-import io.shiftleft.dataflowengineoss.semanticsloader.Semantics
+import io.shiftleft.fuzzyc2cpg.testfixtures.DataFlowCodeToCpgSuite
 import io.shiftleft.semanticcpg.language._
 import overflowdb.traversal.Traversal
+import io.shiftleft.dataflowengineoss.language._
+import io.shiftleft.dataflowengineoss.semanticsloader.Semantics
 
 class CDataFlowTests1 extends DataFlowCodeToCpgSuite {
