+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/querying/DataFlowTests.scala
@@ -1,6 +1,8 @@
-package io.shiftleft.dataflowengineoss.language
+package io.shiftleft.fuzzyc2cpg.querying
 
+import io.shiftleft.fuzzyc2cpg.testfixtures.DataFlowCodeToCpgSuite
 import io.shiftleft.semanticcpg.language._
+import io.shiftleft.dataflowengineoss.language._
 
 class DataFlowTests extends DataFlowCodeToCpgSuite {
   override val code =
