+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/querying/DumpCpg14Tests.scala
@@ -1,7 +1,8 @@
-package io.shiftleft.dataflowengineoss.layers.dataflows
+package io.shiftleft.fuzzyc2cpg.querying
 
 import better.files.File
-import io.shiftleft.dataflowengineoss.language.DataFlowCodeToCpgSuite
+import io.shiftleft.dataflowengineoss.layers.dataflows.{Cpg14DumpOptions, DumpCpg14}
+import io.shiftleft.fuzzyc2cpg.testfixtures.DataFlowCodeToCpgSuite
 import io.shiftleft.semanticcpg.layers.LayerCreatorContext
 
 class DumpCpg14Tests extends DataFlowCodeToCpgSuite {
