+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/querying/TrackingPointTests.scala
@@ -1,7 +1,10 @@
-package io.shiftleft.dataflowengineoss.language
+package io.shiftleft.fuzzyc2cpg.querying
+
 import io.shiftleft.codepropertygraph.generated.nodes
-import io.shiftleft.semanticcpg.language._
 import io.shiftleft.dataflowengineoss.semanticsloader.Semantics
+import io.shiftleft.fuzzyc2cpg.testfixtures.DataFlowCodeToCpgSuite
+import io.shiftleft.semanticcpg.language._
+import io.shiftleft.dataflowengineoss.language._
 
 class TrackingPointTests extends DataFlowCodeToCpgSuite {
