+++ queries/build.sbt
@@ -2,7 +2,8 @@ name := "queries"
 
 dependsOn(Projects.semanticcpg % "compile -> compile; test -> test",
           Projects.dataflowengineoss % "compile -> compile; test -> test",
-          Projects.dataflowengineosstests % "compile -> compile; test -> test"
+          Projects.dataflowengineosstests % "compile -> compile; test -> test",
+          Projects.fuzzyc2cpgtests % "test ->test",
 )
 
 libraryDependencies ++= Seq(
