+++ queries/src/test/scala/io/shiftleft/queries/TruncationsTests.scala
@@ -1,6 +1,6 @@
 package io.shiftleft.queries
 
-import io.shiftleft.dataflowengineoss.language.DataFlowCodeToCpgSuite
+import io.shiftleft.fuzzyc2cpg.testfixtures.DataFlowCodeToCpgSuite
 import io.shiftleft.semanticcpg.language._
 
 class TruncationsTests extends DataFlowCodeToCpgSuite {
